requirements.txt
dj-database-url==0.3.0
Django==1.8.13
django-cors-headers==1.1.0
django-nose==1.4.3
django-secure==1.0.1
djangorestframework==3.1.3
djorm-ext-pgfulltext==0.9.3
model-mommy==1.2.6
newrelic==2.66.0.49
nose==1.3.7
numpy==1.9.3
psycopg2==2.5.5
pytest-django==2.9.1
selenium==2.53.5
six==1.10.0
waitress==0.8.10
whitenoise==3.2
